api/pyproject.toml
[project]
name = "goapp-api"
version = "0.1.0"
description = "Go problem workbook backend: PDF ingestion, stone detection, model training."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "pydantic>=2.9",
  "python-multipart>=0.0.12",
  "pillow>=11.0",
  "numpy>=2.1",
  "pypdfium2>=4.30",
  "opencv-python-headless>=4.10",
]

[project.optional-dependencies]
ml = [
  "torch>=2.5",
  "torchvision>=0.20",
  "onnx>=1.17",
  "onnxruntime>=1.20",
  "ultralytics>=8.3",
]
dev = [
  "pytest>=8",
  "httpx>=0.28",
  "ruff>=0.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/goapp_api"]

[tool.ruff]
line-length = 100
target-version = "py311"
